refactor(types): extract user tests to sibling file
